backend/requirements.txt
# SafeCircle backend — Python 3.11+ (3.12 / 3.13 supported)
# Install: pip install -r requirements.txt

fastapi>=0.115.0,<1.0
uvicorn[standard]>=0.30.0,<1.0
supabase>=2.10.0,<3.0
anthropic>=0.40.0,<1.0
python-dotenv>=1.0.0,<2.0
pydantic>=2.9.0,<3.0
httpx>=0.27.0,<1.0

# Dev / tests
pytest>=8.3.0,<10.0
pytest-asyncio>=0.24.0,<1.0
